Lock with or without args now. Fixes gh-1
With the HASHARGS4LOCK variable, users can now select if the locks should be for unique arguments to the script or if the script should only be locked if the arguments are the same for each instance of the script.  Arguments are now output to the .running file to help identify the process better.